% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jucesp.R
\name{jucesp_parse}
\alias{jucesp_parse}
\title{Parseia ficha cadastral JUCESP em tibble}
\usage{
jucesp_parse(arquivo)
}
\arguments{
\item{arquivo}{Caminho do arquivo HTML}
}
\value{
tibble com uma linha e colunas: \code{cnpj}, \code{nire}, \code{tipo}, \code{data_constituicao},
\code{data_atividade}, \code{inscricao}, \code{objeto}, \code{capital}, \code{logradouro}, \code{numero},
\code{bairro}, \code{complemento}, \code{municipio}, \code{cep}, \code{uf}
}
\description{
Lê o HTML da ficha cadastral salvo por \code{\link[=jucesp_download]{jucesp_download()}}, extrai os
campos da empresa e apaga o arquivo com \code{\link[=unlink]{unlink()}}.
}
